Fix issues that arised in PadListActivity due to nested RecyclerViews.
A selection block system was added to block ActionModes from mixing.
Selection predicates where defined to check the ActionMode block.
An OnTouchListener is created now in the padgroups adapter and then
passed down to the pads adapter to allow the touch events to pass
down to the pad Views so they can be selected when inside a padgroup.